third_party/blink/web_tests/external/wpt/html/browsers/the-window-object/open-close/open-features-non-integer-left-expected.txt

This is a testharness.js-based test.
PASS features "left=/104" should NOT set "left=104"
PASS features "left=_104" should NOT set "left=104"
PASS features "left=L104" should NOT set "left=104"
FAIL features "left=105.5" should set "left=105" assert_equals: "width=401,height=204,top=0,left=105.5 value after first non-digit will be ignored" expected 105 but got 0
FAIL features "left=105.32" should set "left=105" assert_equals: "width=401,height=204,top=0,left=105.32 value after first non-digit will be ignored" expected 105 but got 0
FAIL features "left=105LLl" should set "left=105" assert_equals: "width=401,height=204,top=0,left=105LLl value after first non-digit will be ignored" expected 105 but got 0
FAIL features "left=105^4" should set "left=105" assert_equals: "width=401,height=204,top=0,left=105^4 value after first non-digit will be ignored" expected 105 but got 0
FAIL features "left=105*3" should set "left=105" assert_equals: "width=401,height=204,top=0,left=105*3 value after first non-digit will be ignored" expected 105 but got 0
FAIL features "left=105/5" should set "left=105" assert_equals: "width=401,height=204,top=0,left=105/5 value after first non-digit will be ignored" expected 105 but got 0
FAIL features "left=105  " should set "left=105" assert_equals: "width=401,height=204,top=0,left=105   value after first non-digit will be ignored" expected 105 but got 0
FAIL features "left=105e1" should set "left=105" assert_equals: "width=401,height=204,top=0,left=105e1 value after first non-digit will be ignored" expected 105 but got 0
FAIL features "left=105e-1" should set "left=105" assert_equals: "width=401,height=204,top=0,left=105e-1 value after first non-digit will be ignored" expected 105 but got 0
Harness: the test ran to completion.
